Short-list just the Harfbuzz sources we need
The GNU Make, Cmake, Python scripts, and other random stuff in
Harfbuzz's source directory are not relevant to the cython build which
only uses the amalgam file (harfbuzz.cc) and compiles it using a c++
compiler directly. This includes the other sources and header files, but
doesn't need any of the build machinery.