Fix Windows abi3audit by adding explicit copy command
this fixes the following cibuildwheel error:

https://github.com/harfbuzz/uharfbuzz/actions/runs/18593853794/job/53014973393#step:5:1092

```
Build failed because the repair step completed successfully but
did not produce a wheel.

Your `repair-wheel-command` is expected to place the repaired
wheel in the {dest_dir} directory. See the documentation for
example configurations:

https://cibuildwheel.pypa.io/en/stable/options/#repair-wheel-command
```

Windows has no default repair-wheel-command, so when we set one
(even via inherit+append), we must explicitly copy the wheel to
dest_dir...